requirements.txt
speech_recognitio
pyttsx3
pywhatkit
datetime
wikipedia
pyjokes
webbrowser